+++ src/card.ts
@@ -1,12 +1,4 @@
 export {
-  Button,
-  type ButtonProps,
-  type ButtonShape,
-  type ButtonSize,
-  type ButtonTone,
-  type ButtonVariant,
-} from "./components/button/index.js";
-export {
   Card,
   type CardActionProps,
   type CardContentProps,